chore: move worker to seperate file
