+++ packages/monaco-graphql/src/graphqlMode.ts
@@ -10,7 +10,7 @@ import * as monaco from 'monaco-editor';
 import IRichLanguageConfiguration = monaco.languages.LanguageConfiguration;
 
 import { WorkerManager } from './workerManager';
-import { GraphQLWorker } from './graphql.worker';
+import { GraphQLWorker } from './GraphQLWorker';
 import { monarchLanguage } from './monaco.contribution';
 import { LanguageServiceDefaultsImpl } from './defaults';
 import * as languageFeatures from './languageFeatures';
