+++ packages/monaco-graphql/src/typings/monaco.d.ts
@@ -4,6 +4,7 @@
  *  This source code is licensed under the MIT license found in the
  *  LICENSE file in the root directory of this source tree.
  */
+
 declare module monaco.languages.graphql {
   export interface IDisposable {
     dispose(): void;
@@ -99,6 +100,12 @@ declare module monaco.languages.graphql {
      * Defines whether the built-in selection range provider is enabled.
      */
     readonly selectionRanges?: boolean;
+  }
+
+  export interface ICreateData {
+    languageId: string;
+    enableSchemaRequest: boolean;
+    schemaUrl: string;
   }
 
   export interface LanguageServiceDefaults {
